requirements.txt
base58==2.1.1
click==8.0.4
colorama==0.4.4
greenlet==1.1.2
protobuf==3.19.4
protoletariat==0.7.5
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
rocksdict==0.2.16
SQLAlchemy==1.4.31
typing-extensions==4.1.1
